// client/src/types/links.ts
export enum ESelectedPage {
  home = "home",
  shop = "shop",
  categories = "categories",
  about = "about",
  FAQ = "FAQ",
}

export enum ERoutes {
  home = "/",
  martens = "/martens",
  authpage = "/authpage",
  verify = "/verify",
  footwear = "/footwear",
  category = "/category",
  resetpassword = "/reset-password",
  useraccount = "/useraccount",
  shipping = "/shipping",
  payment = "/payment",
  order = "/order",
  userorders = "/userorders",
  createorder = "/createorder",
}